+++ src/App.js
@@ -3,34 +3,13 @@ import "katex/dist/katex.min.css";
 import TeX from "@matejmazur/react-katex";
 import script from "./python/main.py";
 import "./App.css";
-import { CxPlot } from "./CxPlot.js";
+import { CxPlot } from "./components/CxPlot.js";
+import { PyodideButton } from "./components/PyodideButton";
 import { create, all } from "mathjs";
 import { TextField, Grid, Box, Typography } from "@mui/material";
-import { LoadingButton } from "@mui/lab";
 
-const pyodideURL = "https://cdn.jsdelivr.net/pyodide/v0.21.3/full/";
 const math = create(all);
 const nerdamer = require("nerdamer");
-
-function loadScript(url) {
-  return new Promise(function (resolve, reject) {
-    let script = document.createElement("script");
-    script.src = url;
-    script.onload = resolve;
-    script.onerror = reject;
-    document.head.appendChild(script);
-  });
-}
-
-export const loadPyodide = async () => {
-  await loadScript(pyodideURL + "pyodide.js");
-  const pyodide = await window.loadPyodide({
-    indexURL: pyodideURL,
-  });
-  // Packages in https://github.com/pyodide/pyodide/tree/main/packages
-  await pyodide.loadPackage(["micropip", "numpy", "scipy", "sympy"]);
-  return pyodide;
-};
 
 export const runCxroots = async (pyodide, pythonArgs) => {
   const scriptText = await (await fetch(script)).text();
@@ -39,40 +18,6 @@ export const runCxroots = async (pyodide, pythonArgs) => {
   window.self["cxroots_args"] = pythonArgs;
 
   return await pyodide.runPythonAsync(scriptText);
-};
-
-/**
- * PyodideButton loads pyodide before being clickable and does
- * not allow the button to be clicked while pyodide is running
- */
-export const PyodideButton = ({ disabled, onClick }) => {
-  const [pyodide, setPyodide] = useState(null);
-  const [loading, setLoading] = useState(false);
-
-  if (pyodide == null) {
-    setPyodide("loading");
-    setLoading(true);
-    loadPyodide()
-      .then(setPyodide)
-      .then(() => setLoading(false));
-  }
-
-  const onClickLoading = async (event) => {
-    setLoading(true);
-    await onClick(event, pyodide);
-    setLoading(false);
-  };
-
-  return (
-    <LoadingButton
-      loading={loading}
-      disabled={disabled}
-      variant="contained"
-      onClick={onClickLoading}
-    >
-      Find the Roots
-    </LoadingButton>
-  );
 };
 
 const ParseLatex = (text) => {
